# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Memoh API",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/memory/add": {
            "post": {
                "description": "Add memory for a user via memory. Auth: Bearer JWT determines user_id (sub or user_id).",
                "tags": [
                    "memory"
                ],
                "summary": "Add memory",
                "parameters": [
                    {
                        "description": "Add request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.memoryAddPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/memory.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/memory/embed": {
            "post": {
                "description": "Embed text or multimodal input and upsert into memory store. Auth: Bearer JWT determines user_id (sub or user_id).",
                "tags": [
                    "memory"
                ],
                "summary": "Embed and upsert memory",
                "parameters": [
                    {
                        "description": "Embed upsert request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.memoryEmbedUpsertPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/memory.EmbedUpsertResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/memory/memories": {
            "get": {
                "description": "List memories for a user via memory. Auth: Bearer JWT determines user_id (sub or user_id).",
                "tags": [
                    "memory"
                ],
                "summary": "List memories",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run ID",
                        "name": "run_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/memory.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete all memories for a user via memory. Auth: Bearer JWT determines user_id (sub or user_id).",
                "tags": [
                    "memory"
                ],
                "summary": "Delete memories",
                "parameters": [
                    {
                        "description": "Delete all request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.memoryDeleteAllPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/memory.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/memory/memories/{memoryId}": {
            "get": {
                "description": "Get a memory by ID via memory. Auth: Bearer JWT determines user_id (sub or user_id).",
                "tags": [
                    "memory"
                ],
                "summary": "Get memory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Memory ID",
                        "name": "memoryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/memory.MemoryItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a memory by ID via memory. Auth: Bearer JWT determines user_id (sub or user_id).",
                "tags": [
                    "memory"
                ],
                "summary": "Delete memory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Memory ID",
                        "name": "memoryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/memory.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/memory/search": {
            "post": {
                "description": "Search memories for a user via memory. Auth: Bearer JWT determines user_id (sub or user_id).",
                "tags": [
                    "memory"
                ],
                "summary": "Search memories",
                "parameters": [
                    {
                        "description": "Search request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.memorySearchPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/memory.SearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/memory/update": {
            "post": {
                "description": "Update a memory by ID via memory. Auth: Bearer JWT determines user_id (sub or user_id).",
                "tags": [
                    "memory"
                ],
                "summary": "Update memory",
                "parameters": [
                    {
                        "description": "Update request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/memory.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/memory.MemoryItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.memoryAddPayload": {
            "type": "object",
            "properties": {
                "embedding_enabled": {
                    "type": "boolean"
                },
                "filters": {
                    "type": "object",
                    "additionalProperties": true
                },
                "infer": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/memory.Message"
                    }
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "run_id": {
                    "type": "string"
                }
            }
        },
        "handlers.memoryDeleteAllPayload": {
            "type": "object",
            "properties": {
                "run_id": {
                    "type": "string"
                }
            }
        },
        "handlers.memoryEmbedUpsertPayload": {
            "type": "object",
            "properties": {
                "filters": {
                    "type": "object",
                    "additionalProperties": true
                },
                "input": {
                    "$ref": "#/definitions/memory.EmbedInput"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "model": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "run_id": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "handlers.memorySearchPayload": {
            "type": "object",
            "properties": {
                "embedding_enabled": {
                    "type": "boolean"
                },
                "filters": {
                    "type": "object",
                    "additionalProperties": true
                },
                "limit": {
                    "type": "integer"
                },
                "query": {
                    "type": "string"
                },
                "run_id": {
                    "type": "string"
                },
                "sources": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "memory.DeleteResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "memory.EmbedInput": {
            "type": "object",
            "properties": {
                "image_url": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
        "memory.EmbedUpsertResponse": {
            "type": "object",
            "properties": {
                "dimensions": {
                    "type": "integer"
                },
                "item": {
                    "$ref": "#/definitions/memory.MemoryItem"
                },
                "model": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                }
            }
        },
        "memory.MemoryItem": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "memory": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "runId": {
                    "type": "string"
                },
                "score": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "memory.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "memory.SearchResponse": {
            "type": "object",
            "properties": {
                "relations": {
                    "type": "array",
                    "items": {}
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/memory.MemoryItem"
                    }
                }
            }
        },
        "memory.UpdateRequest": {
            "type": "object",
            "properties": {
                "embedding_enabled": {
                    "type": "boolean"
                },
                "memory": {
                    "type": "string"
                },
                "memory_id": {
                    "type": "string"
                }
            }
        },
...
    }
}